ref: refactor plugins/client-resources init
