+++ plugins/client-resources/src/RequestPromise.ts
@@ -1,0 +1,34 @@
+//
+// Copyright © 2020, 2021 Anticrm Platform Contributors.
+// Copyright © 2021 Hardcore Engineering Inc.
+//
+// Licensed under the Eclipse Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License. You may
+// obtain a copy of the License at https://www.eclipse.org/legal/epl-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+export default class RequestPromise {
+  readonly promise: Promise<any>
+  resolve!: (value?: any) => void
+  reject!: (reason?: any) => void
+  reconnect?: () => void
+  constructor (
+    readonly method: string,
+    readonly params: any[],
+    readonly handleResult?: (result: any) => Promise<void>
+  ) {
+    this.promise = new Promise((resolve, reject) => {
+      this.resolve = resolve
+      this.reject = reject
+    })
+  }
+
+  chunks?: { index: number, data: any[] }[]
+}
